Refactor(#72): 폴더 구조 변경2 및 파일 이름 수정
