+++ src/auth/auth.module.ts
@@ -5,14 +5,14 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { AuthSocialLoginRepository } from './repositories/auth-social-login.repository';
 import { UserRepository } from 'src/user/repositories/user.repository';
 import { AuthPasswordLoginRepository } from './repositories/auth-password-login.repository';
-import { UserProfileRepository } from 'src/user/repositories/userProfile.repository';
+import { UserProfileRepository } from 'src/user/repositories/user-profile.repository';
 import { PassportModule } from '@nestjs/passport';
 import { JwtModule } from '@nestjs/jwt';
 import * as config from 'config';
 import { JwtStrategy } from '../config/strategis/jwt-access-token.strategy';
-import { RedisModule } from 'src/redis/redis.module';
+import { RedisModule } from 'src/common/redis/redis.module';
 import JwtRefreshStrategy from '../config/strategis/jwt-refresh-token.strategy';
-import { UserImageRepository } from 'src/user/repositories/userImage.repository';
+import { UserImageRepository } from 'src/user/repositories/user-image.repository';
 import { AuthSocialService } from './services/auth-social.service';
 import { AuthSocialController } from './controllers/auth-social.controller';
