+++ src/chats/controllers/lightning-chats.controller.ts
@@ -20,7 +20,7 @@ import { CreateLightningChattingDto } from '../dtos/create-lightning-chattings.d
 import { HttpExceptionFilter } from 'src/common/exceptions/http-exception.filter';
 import { SuccessInterceptor } from 'src/common/interceptors/success.interceptor';
 import { JwtAccessGuard } from 'src/config/guards/jwt-access-token.guard';
-import { GetUserId } from 'src/common/decorators/getUserId.decorator';
+import { GetUserId } from 'src/common/decorators/get-userId.decorator';
 
 @ApiTags('lightning-chats')
 @Controller('lightning-chats')
