+++ src/config/swagger.ts
@@ -1,6 +1,6 @@
 import { INestApplication } from '@nestjs/common';
 import { DocumentBuilder, SwaggerModule } from '@nestjs/swagger';
-import expressBasicAuth from 'express-basic-auth';
+import * as expressBasicAuth from 'express-basic-auth';
 import * as config from 'config';
 
 const swaggerConfig = config.get('swagger');
@@ -15,6 +15,7 @@ export function setupSwagger(app: INestApplication): void {
       },
     }),
   );
+
   const option = new DocumentBuilder()
     .setTitle('5term-main-api')
     .setDescription('모임서비스API')
@@ -31,5 +32,5 @@ export function setupSwagger(app: INestApplication): void {
     .build();
 
   const document = SwaggerModule.createDocument(app, option);
-  SwaggerModule.setup('api-docs', app, document);
+  SwaggerModule.setup('docs', app, document);
 }
