+++ src/lightning/controllers/lightning.controller.ts
@@ -17,7 +17,7 @@ import { UpdateLightningToUserDto } from '../dtos/update-lightning-to-user.dto';
 import { RequestLightningDto } from '../dtos/request-lightning.dto';
 import { UpdateAcceptLightningDto } from '../dtos/update-accept-lightning.dto';
 import { JwtAccessGuard } from 'src/config/guards/jwt-access-token.guard';
-import { GetUserId } from 'src/common/decorators/getUserId.decorator';
+import { GetUserId } from 'src/common/decorators/get-userId.decorator';
 
 @ApiTags('lightning')
 @Controller('lightnings')
