+++ src/study_recruit/controller/study_recruit.controller.ts
@@ -16,9 +16,9 @@ import { HttpExceptionFilter } from 'src/common/exceptions/http-exception.filter
 import { SuccessInterceptor } from 'src/common/interceptors/success.interceptor';
 import { StudyRecruitService } from '../service/study_recruit.service';
 import { JwtAccessGuard } from 'src/config/guards/jwt-access-token.guard';
-import { GetUserId } from 'src/common/decorators/getUserId.decorator';
+import { GetUserId } from 'src/common/decorators/get-userId.decorator';
 import { FilesInterceptor } from '@nestjs/platform-express';
-import { S3Service } from 'src/s3/s3.service';
+import { S3Service } from 'src/common/s3/s3.service';
 
 @ApiTags('study-recruitment')
 @Controller('study-recruit')
