+++ src/study_recruit/study_recruit.module.ts
@@ -4,7 +4,7 @@ import { StudyRecruitService } from './service/study_recruit.service';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { StudyRecruitBoardRepository } from './reoisitories/study_recruitment_repository';
 import { StudyAdminsRepository } from 'src/study/repository/study_admins.repository';
-import { S3Service } from 'src/s3/s3.service';
+import { S3Service } from 'src/common/s3/s3.service';
 import { StudyRecruitBoardImgRepository } from './reoisitories/study_recruitment_IMG_repository';
 
 @Module({
