+++ src/user/repositories/user-profile.repository.ts
@@ -1,11 +1,11 @@
 import { EntityRepository, Repository } from 'typeorm';
-import { UserProfile } from '../entities/user_profile.entity';
+import { UserProfile } from '../entities/user-profile.entity';
 import { AuthCredentialDto } from '../../auth/dtos/auth-credential.dto';
 import { User } from '../entities/user.entity';
 import { NicknameDuplicationCheckDto } from 'src/auth/dtos/duplication-check.dto';
 import { BadRequestException } from '@nestjs/common';
 import { SocialUserProfileDto } from 'src/auth/dtos/social-user-profile.dto';
-import { UserImage } from '../entities/user_image.entity';
+import { UserImage } from '../entities/user-image.entity';
 
 @EntityRepository(UserProfile)
 export class UserProfileRepository extends Repository<UserProfile> {
