+++ src/user/types/schedule.enum.ts
